+++ specs/client.js
@@ -1,0 +1,8 @@
+import './client/exports';
+import './client/apollo-link-ddp';
+import './client/apollo-client';
+import './client/apollo-link-meteor-auth';
+import './client/apollo-link-meteor';
+import './client/graphQLFetcher';
+import './client/asteroid';
+import './client/simpleddp';