+++ specs/client.spec.js
@@ -1,6 +1,0 @@
-import './client/exports.spec';
-import './client/apollo-link-ddp.spec';
-import './client/apollo-client.spec';
-import './client/apollo-link-meteor-auth.spec';
-import './client/apollo-link-meteor.spec';
-import './client/graphQLFetcher.spec';