+++ specs/client/apollo-link-ddp.js
@@ -3,10 +3,15 @@
 import chai from 'chai';
 import gql from 'graphql-tag';
 import { ApolloLink, Observable } from 'apollo-link';
+import {
+  DEFAULT_METHOD,
+  DEFAULT_PUBLICATION,
+  getDDPLink,
+  DDPMethodLink,
+  DDPSubscriptionLink,
+} from 'apollo-link-ddp';
 import { loginWithUserId } from './helpers/login';
 import { callPromise } from './helpers/callPromise';
-import { getDDPLink, DDPMethodLink, DDPSubscriptionLink } from '../../lib/client/apollo-link-ddp';
-import { DEFAULT_METHOD, DEFAULT_PUBLICATION, GRAPHQL_SUBSCRIPTION_MESSAGE_TYPE } from '../../lib/common/defaults';
 import { FOO_CHANGED_TOPIC } from '../data/resolvers';
 
 describe('DDPMethodLink', function () {
@@ -95,6 +100,27 @@ describe('DDPMethodLink', function () {
           try {
             chai.expect(data.meteorUserId).to.be.a('string');
             chai.expect(data.meteorUserId).to.equal(this.userId);
+            done();
+          } catch (e) {
+            done(e);
+          }
+        },
+        error: done,
+      });
+    });
+
+    it('returns the isDDP flag', function (done) {
+      const operation = {
+        query: gql`query { isDDP }`,
+      };
+
+      const observer = this.link.request(operation);
+
+      observer.subscribe({
+        next: ({ data }) => {
+          try {
+            chai.expect(data.isDDP).to.be.a('boolean');
+            chai.expect(data.isDDP).to.be.true;
             done();
           } catch (e) {
             done(e);
@@ -232,45 +258,6 @@ describe('DDPSubscriptionLink', function () {
             done(e);
           }
         }, 100);
-      });
-    });
-
-    it('accepts a custom DDP observer', function (done) {
-      const operation = {
-        query: gql`subscription { fooSub }`,
-      };
-      const message = { fooSub: 'custom' };
-      let customObserverLink;
-
-      const ddpObserver = new Observable((observer) => {
-        setTimeout(() => {
-          observer.next({
-            type: GRAPHQL_SUBSCRIPTION_MESSAGE_TYPE,
-            subId: customObserverLink.subscriptionObservers.keys().next().value,
-            graphqlData: { data: { ...message } },
-          });
-          observer.complete();
-        }, 10);
-      });
-
-      customObserverLink = new DDPSubscriptionLink({ ddpObserver });
-
-      chai.expect(customObserverLink.ddpObserver).to.equal(ddpObserver);
-
-      const observer = customObserverLink.request(operation);
-
-      const subscription = observer.subscribe({
-        next: ({ data }) => {
-          try {
-            chai.expect(data).to.deep.equal(message);
-            subscription.unsubscribe();
-            customObserverLink.ddpSubscription.unsubscribe();
-            done();
-          } catch (e) {
-            done(e);
-          }
-        },
-        error: done,
       });
     });
   });