+++ specs/server.js
@@ -1,0 +1,8 @@
+// helpers
+import './server/helpers';
+import './server/helpers/setupLoginByUserId';
+
+// specs
+import './server/server';
+import './server/setup';
+import './server/getUserIdByLoginToken';